Moody's/Extract Base Pricing/requirements.txt
playwright
pyyaml
